refactor: 재사용할 수 있는 컴포넌트 common 폴더로 이동
